% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_transactions_data_for_adwords.r
\name{clean_transactions_data_for_adwords}
\alias{clean_transactions_data_for_adwords}
\title{Clean Raw Transaction Data from Application Database}
\usage{
clean_transactions_data_for_adwords(data_frame)
}
\arguments{
\item{data_frame}{A data frame of transaction events from \code{get_transactions_data}}
}
\value{
A data frame with the transaction events cleaned and consistent with other data frames.
}
\description{
Cleans the transactions data from the App database for AdWords
}
\examples{
clean_transactions_data(data_frame)
}

