% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bing_event_log.r
\name{create_bing_event_log}
\alias{create_bing_event_log}
\title{Create Keywords Event Log (elog) for Bing}
\usage{
create_bing_event_log(from = Sys.Date(), to = Sys.Date())
}
\arguments{
\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\value{
A data frame with all events (Bing clicks, Transactions and Referrals).
}
\description{
Created an event log (elog) of all transactions, referrals and keyword activity.  An event log
is simple a time-series log of events that happen.  When a certain row (observation) does not include
a particular variable, that variable is marked as NA.  For example, a Bing keyword observation will
have information about the number of clicks, but no revenue for the transaction.  Revenue data, instead
is pulled in from the order application's data.
}
\examples{
create_bing_event_log(from=20150101, to=20151231)
}

