% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_google_display_event_log.r
\name{create_google_display_event_log}
\alias{create_google_display_event_log}
\title{Create Google Display Event Log (elog) for AdWords}
\usage{
create_google_display_event_log(from = Sys.Date(), to = Sys.Date())
}
\arguments{
\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\value{
A data frame with all events (AdWords clicks, Transactions and Referrals).
}
\description{
Create Google Display Event Log (elog) for AdWords
}
\examples{
create_google_display_event_log(from=20150101, to=20151231)
}

