% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first_transaction_dates_for_users.R
\name{get_first_transaction_dates_for_users}
\alias{get_first_transaction_dates_for_users}
\title{Get First Transactions Dates for Users}
\usage{
get_first_transaction_dates_for_users(from = Sys.Date(), to = Sys.Date(),
  users = NULL,
  database_driver = "database_drivers/postgresql-9.4.1208.jre6.jar",
  jdbc_config_file = "jdbc_heroku_string.txt")
}
\arguments{
\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{users}{A vector of user IDs that should be in the database.}

\item{database_driver}{Path to the database driver for the database being used. Defaults to "database_drivers/postgresql-9.4.1208.jre6.jar".}

\item{jdbc_config_file}{Path to a text file with the URI to connect to the database, including credentials. Defaults to "jdbc_data_warehouse_database_config.txt"}

\item{transactions_query_file}{Path to a .sql file that includes the SQL query to be used to retrieve the necessary data from the database. Defaults to "referrals_query.sql"}
}
\value{
A data frame with transaction data within the date range.
}
\description{
Using a vector of user IDs, this function returns a data frame with two columns: 
the user's ID and the date of their first transaction if their first transaction
falls within the from and to dates.  If their first transaction falls outside of those dates.
the user's information is not included.
}
\examples{
get_first_transaction_dates_for_users(from=20150101, to=20151231, users=transactions$users)
}

