% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transactions_data.r
\name{get_transactions_data}
\alias{get_transactions_data}
\title{Get Transactions Data}
\usage{
get_transactions_data(from = Sys.Date(), to = Sys.Date(),
  database_driver = "database_drivers/postgresql-9.4.1208.jre6.jar",
  jdbc_config_file = "jdbc_transactions_database_config.txt",
  transactions_query_file = "transactions_query.sql")
}
\arguments{
\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{database_driver}{Path to the database driver for the database being used. Defaults to "database_drivers/postgresql-9.4.1208.jre6.jar".}

\item{jdbc_config_file}{Path to a text file with the URI to connect to the database, including credentials. Defaults to "jdbc_transactions_database_config.txt"}

\item{transactions_query_file}{Path to a .sql file that includes the SQL query to be used to retrieve the necessary data from the database. Defaults to "transactions_query.sql"}
}
\value{
A data frame with transaction data within the date range.
}
\description{
Returns a data frame with all transaction data within a date range.  Data is 
pulled from real time data
}
\examples{
get_transactions_data(from=20150101, to=20151231)
}

