% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_per_week_by_geo.r
\name{orders_per_week}
\alias{orders_per_week}
\title{Create plot of number of orders per week by geo}
\usage{
orders_per_week(keywords_elog, plot = TRUE)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}
}
\value{
plot logical. If TRUE, creates a plot of the data.  If false, creates no plot, Defaults to \code{TRUE}.

The data frame used to create a plot of performance over time.
}
\description{
Takes in a AdWords event log and processes all summary data
}
\examples{
overall_performance_over_time(elog_data_frame)
}

