% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleo_cohort_views.r
\name{paleo_cohort_views}
\alias{paleo_cohort_views}
\title{Create plot of cohorts who converted from the keyword "Paleo Meals"}
\usage{
paleo_cohort_views(keywords_elog)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}
}
\description{
Takes in an AdWords or Bing event log.
}
\examples{
paleo_cohort_views(keywords_elog)
}

