% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_overview.r
\name{user_overview}
\alias{user_overview}
\title{Generate User Performance Overview}
\usage{
user_overview(keywords_elog)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}
}
\value{
The data frame used to create a plot of performance over time.
}
\description{
Takes in a AdWords or Bing event log and creates a revenue breakdown by users
}

