% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transactions_data.r
\name{get_transactions_data}
\alias{get_transactions_data}
\title{Get Transactions Data}
\usage{
get_transactions_data(from = Sys.Date(), to = Sys.Date(),
  database_driver = "database_drivers/postgresql-9.4.1208.jre6.jar",
  jdbc_config_file = "jdbc_transactions_database_config.txt",
  transactions_query_file = "transactions_query.sql")
}
\arguments{
\item{from}{Start date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}

\item{to}{End date in either format <"yyyy-mm-dd"> or <yyyymmdd>. Inclusive.}
}
\value{
A data frame with transaction data within the date range.
}
\description{
Returns a data frame with all transaction data within a date range.  Data is 
pulled from real time data
}
\examples{
get_transactions_data(from=20150101, to=20151231)
}

