% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orders_per_week.r
\name{orders_per_week}
\alias{orders_per_week}
\title{Create plot of number of orders per week by geo}
\usage{
orders_per_week(keywords_elog, keyword_filter = NULL,
  campaign_filter = NULL, plot = TRUE)
}
\arguments{
\item{keywords_elog}{A data frame containing keywords and transaction event log data.}

\item{keyword_filter}{Optional. (e.g. "paleo meals") If provided, generates a data frame and plot only for the *exact* keyword provided. Default to NULL.}

\item{campaign_filter}{Optional. (e.g. "Paleo Performers") If provided, generates a data frame and plot for campaigns where the string is included in the campaign name. Default to NULL.}
}
\value{
plot logical. If TRUE, creates a plot of the data.  If false, creates no plot, Defaults to \code{TRUE}.

The data frame used to create a plot of performance over time.
}
\description{
Takes in a AdWords event log and processes all summary data
}
\examples{
orders_per_week(keywords_elog, keyword_filter="paleo meals", campaign_filter="Paleo Performers")
}

